requirements.txt
Django==5.0
djangorestframework==3.14.0
langchain==0.2.4
langchain_core==0.2.6
langsmith==0.1.77
openai==1.34.0
pinecone_client==2.2.4
pydantic==1.10.9
pyhumps==3.8.0
python-dotenv==1.0.1
tqdm==4.66.2
